feat: Add comments and tags to posts
This feature allows users to add a comment and custom tags when posting a 'Now Playing' track. - Adds a textarea for comments and an input for tags in thepopup UI. - The #nowplaying tag is added by default. - Updates the background script to include the comment and tags in the Mastodon post.